% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeAdjustedSelectionUsingRecomputeMethod}
\alias{computeAdjustedSelectionUsingRecomputeMethod}
\title{Compute selection - recompute method}
\usage{
computeAdjustedSelectionUsingRecomputeMethod(dist, adjustedSelected)
}
\arguments{
\item{dist}{distance matrix, used for distances and implicitly defining the set of entries}

\item{adjustedSelected}{the selected entries defined as a list of zero-based integers referring to the row/columns of \code{dist}}
}
\description{
The function \code{computeAdjustedSelectionUsingRecomputeMethod} is used internally by
the \code{\link{CoreCollection}} object to compute an adjusted selection using the recompute
method.
}
\details{
This function returns a list describing for each of the row/columns entries of \code{dist} the
closest selected entry. The entries are implicetly defined by the row/columns of
\code{dist} and referred to by a zero-based integer describing the position.
}
\keyword{internal}
