% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreCollection.R
\docType{class}
\name{.CoreCollectionClass}
\alias{.CoreCollectionClass}
\title{The CoreCollection Class}
\format{A \code{\link{R6Class}} generator object}
\usage{
.CoreCollectionClass
}
\description{
The CoreCollection Class
}
\keyword{internal}
