% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreCollection.package.R
\docType{package}
\name{coreCollection-package}
\alias{coreCollection-package}
\alias{coreCollection}
\title{The coreCollection package}
\description{
This package can be used to create a \link{CoreCollection} object.
}
\references{
Odong, T.L. (2012) Quantative methods for sampling of germplasm collections -
Getting the best out of molecular markers when creating core collections. PhD diss., Wageningen
University and Research, Wageningen, The Netherlands. \url{http://edepot.wur.nl/212422}\cr

Jansen, J & Hintum, Theo. (2007) Genetic distance sampling: A novel
sampling method for obtaining core collections using genetic distances with an
application to cultivated lettuce. TAG. Theoretical and applied genetics.
Theoretische und angewandte Genetik. 114. 421-8. 10.1007/s00122-006-0433-9\cr
}
\seealso{
- \pkg{vcfR} provides a suite of tools for input and output of variant call format (VCF) files, manipulation of their content and visualization.\cr
- \pkg{adegenet} provides the \code{genlight} class for genome-wide SNP data, and includes a method to create a distance matrix.

Other core collection: 
\code{\link{CoreCollection}()}
}
\author{
Matthijs Brouwer <matthijs.brouwer@wur.nl>
}
\concept{core collection}
