% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreCollection.R
\docType{class}
\name{.CoreCollectionClass}
\alias{.CoreCollectionClass}
\title{The CoreCollection Class}
\format{
A \code{\link{R6Class}} generator object
}
\description{
The CoreCollection Class

The CoreCollection Class
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-coreCollection-recompute}{\code{.CoreCollectionClass$recompute()}}
\item \href{#method-coreCollection-alternativeCore}{\code{.CoreCollectionClass$alternativeCore()}}
\item \href{#method-coreCollection-new}{\code{.CoreCollectionClass$new()}}
\item \href{#method-coreCollection-print}{\code{.CoreCollectionClass$print()}}
\item \href{#method-coreCollection-summary}{\code{.CoreCollectionClass$summary()}}
\item \href{#method-coreCollection-measure}{\code{.CoreCollectionClass$measure()}}
\item \href{#method-coreCollection-measures}{\code{.CoreCollectionClass$measures()}}
\item \href{#method-coreCollection-clone}{\code{.CoreCollectionClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coreCollection-recompute"></a>}}
\if{latex}{\out{\hypertarget{method-coreCollection-recompute}{}}}
\subsection{Method \code{recompute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.CoreCollectionClass$recompute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coreCollection-alternativeCore"></a>}}
\if{latex}{\out{\hypertarget{method-coreCollection-alternativeCore}{}}}
\subsection{Method \code{alternativeCore()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.CoreCollectionClass$alternativeCore(n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coreCollection-new"></a>}}
\if{latex}{\out{\hypertarget{method-coreCollection-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.CoreCollectionClass$new(
  distanceMatrix,
  n,
  preselected,
  coreSelectMethod,
  adjustedGroupMethod,
  algorithm,
  seed
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coreCollection-print"></a>}}
\if{latex}{\out{\hypertarget{method-coreCollection-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.CoreCollectionClass$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coreCollection-summary"></a>}}
\if{latex}{\out{\hypertarget{method-coreCollection-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.CoreCollectionClass$summary(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coreCollection-measure"></a>}}
\if{latex}{\out{\hypertarget{method-coreCollection-measure}{}}}
\subsection{Method \code{measure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.CoreCollectionClass$measure(coreSelectMethod)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coreCollection-measures"></a>}}
\if{latex}{\out{\hypertarget{method-coreCollection-measures}{}}}
\subsection{Method \code{measures()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.CoreCollectionClass$measures(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coreCollection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-coreCollection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.CoreCollectionClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
